+++ src/03_semantic/root.zig
@@ -53,6 +53,7 @@ pub fn semantic_analysis_unmanaged(
     }
 
     // Name resolution
+
     // Type checking
     var typechecker_visitor = TypecheckerVisitor.init(alloc, symbol_table, &symbol_table.scope, &semantic_ctx, err);
     const type_visitor = typechecker_visitor.visitor();